+++ src/utils/git.js
@@ -13,7 +13,7 @@ const { buildRemotePayload, filterUniqueRemotes } = require(path.resolve(
  * @param {String} remote Current displayed remote
  * @returns {String[]} payload and uniqueRemotes
  */
-export function getRefData() {
+export const getRefData = () => {
   const refs = getRefs();
 
   const payload = refs
@@ -33,7 +33,7 @@ export function getRefData() {
     });
 
   return [payload, uniqueRemotes];
-}
+};
 
 /**
  * Pull branch information from selection and pass as args to execGit().
@@ -41,7 +41,7 @@ export function getRefData() {
  * Returns a promise that resolves when the user has successfully checked out
  * target branch
  */
-export function doCheckoutBranch(branch, remote) {
+export const doCheckoutBranch = (branch, remote) => {
   let branchPath = '';
 
   if (remote && remote !== 'local' && remote !== 'origin') {
@@ -53,25 +53,25 @@ export function doCheckoutBranch(branch, remote) {
   const args = ['checkout', branchPath];
 
   return execGit(args);
-}
+};
 
 /**
  * Return name of current branch.
  *
  * Returns a promise that resolves to the current branch name.
  */
-export function getCurrentBranch() {
+export const getCurrentBranch = () => {
   const args = ['rev-parse', '--abbrev-ref', 'HEAD'];
 
   return execGit(args);
-}
+};
 
 /**
  * Execute git command with passed arguments.
  * <args> is expected to be an array of strings.
  * Example: ['fetch', '-pv']
  */
-function execGit(args) {
+const execGit = args => {
   return new Promise((resolve, reject) => {
     let dataString = '';
     let errorString = '';
@@ -92,18 +92,18 @@ function execGit(args) {
       }
     });
   });
-}
+};
 
 /**
  * Call git fetch with a prune and quiet flag.
  *
  * Return a promise that resolves when a user successfully fetches.
  */
-export function doFetchBranches() {
+export const doFetchBranches = () => {
   const args = ['fetch', '-pq'];
 
   return execGit(args);
-}
+};
 
 /**
  * Format output from getBranchesFrom() and return an array of arrays containing
@@ -112,7 +112,7 @@ export function doFetchBranches() {
  * @param {String} output String list of each ref associated with repository
  * @return {Array} Array containing an array of line items representing branch information
  */
-export function formatRemoteBranches(output) {
+export const formatRemoteBranches = output => {
   let remoteBranchArray = [];
 
   return getCurrentBranch().then(selectedBranch => {
@@ -144,14 +144,14 @@ export function formatRemoteBranches(output) {
 
     return remoteBranchArray;
   });
-}
+};
 
 /**
  * Print all refs assicated with git repository.
  *
  * @return {String} String list of each ref associated with repository.
  */
-function getRefs() {
+const getRefs = () => {
   const args = [
     'for-each-ref',
     '--sort=-committerdate',
@@ -162,4 +162,4 @@ function getRefs() {
   return execGit(args).then(data => {
     return formatRemoteBranches(data);
   });
-}
+};